requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
redis
pymongo
python-dotenv
playwright
streamlit
